+++ setup.py
@@ -1,31 +1,8 @@
 import os
 import setuptools
-import sys
-import subprocess
 
 package_name = 'alpha-automl'
 package_dir = 'alpha_automl'
-
-
-class get_pybind_include(object):
-    """Helper class to determine the pybind11 include path
-
-    The purpose of this class is to postpone importing pybind11
-    until it is actually installed, so that the ``get_include()``
-    method can be invoked. """
-
-    def __init__(self, user=False):
-        try:
-            import pybind11
-        except ImportError:
-            if subprocess.call([sys.executable, '-m', 'pip', 'install', 'pybind11']):
-                raise RuntimeError('pybind11 install failed.')
-
-        self.user = user
-
-    def __str__(self):
-        import pybind11
-        return pybind11.get_include(self.user)
 
 
 def read_readme():
@@ -71,13 +48,4 @@ setuptools.setup(
         'Intended Audience :: Science/Research',
         'License :: OSI Approved :: Apache Software License',
         'Topic :: Scientific/Engineering',
-    ],
-    ext_modules=[
-        setuptools.Extension(
-            'your_extension_name',
-            sources=['your_extension_source.cpp'],
-            include_dirs=[get_pybind_include()],
-            language='c++'
-        )
-    ]
-)
+    ])
